Moved CLI tests into separate file
